+++ src/spc/steps/postprocess/copyViteCoverage.ts
